requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
influxdb==5.2.2
psycopg2-binary==2.8.2
pymemcache==2.1.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
simple-websocket-server==0.4.0
six==1.12.0
urllib3==1.25.3
websocket-client==0.56.0
